It works!
'juju sync-tools' can now look at the ec2 public bucket, download everything,
and then upload it to the specified environment.
We need to be a bit careful, because we tend to get login credentials
from the environment, but now we are connecting to *2* services. So 
they must be specified elsewhere. But we can make it work.