charm: expose CharmStore.Info and CharmStore.Event
The low-level API of the charm store must be accessible to implement
functionality that is not about a Repository, but rather about the
store itself. This will be used to implement the upcoming "publish"
juju command.

The Repository interface must still be used when one wants to talk
to a repositroy generically.