Create a NotifyWorker
This matches the idea of NotifyWatcher.
It is designed to handle all the Worker logic, so that actual
workers can focus on just doing work.

This handles monitoring the Changes channel, and the Tomb for
requests to shutdown. It also handles stopping the Watcher
when shutting down, error propagation, etc.

This allows us to create new workers based on a NotifyWatcher
with 3 simple setup, handle, and teardown functions.