Rework the internals a bit.
Instead of starting a NotifyWorker with a Watcher already running,
have SetUp create the Watcher (existing workers seemed to all use
that pattern.)
Further, use an interface rather than 3 functions. It makes
sense that almost everything that calls SetUp is going to
be setting internal state, and that state should be used in
Handle, and definitely used in TearDown.
An interface lets all that hang off of the state that it is
managing, rather than always having to wrap your state in
func() calls.