Add tests specific to the MachineAgent.
Discover that machine agents that aren't machine-0 weren't
forcing a state connection, but trying to read the API connection
to determine if they had state jobs to perform. However, they
can't get an API connection on an upgrade from juju-1.10.
So for now, always start the state connection, so that we
can set the API password.