Handle the no-api-password differently.
Instead of always creating the state worker early, wait until
we fail to login to the API, and then force a state connection
at that point.

This gets us closer to what we want (no state connection unless
we need it). It also means once we've set the API password,
we won't have to start the state worker early.