Handle the fact that APIInfo is completely missing in 1.10 unit agent.conf.
We need to copy the contents from StateInfo, and then guess what port the API is on.
This will work for everything but if someone specifies a custom port in
environments.yaml, but I don't know why they would.
Perhaps it could be done with a State lookup for the API info, I'll ask in review.
Note, though, this is mostly compatibility so getting the 99% case should be enough.