uniter: Use abstract namespace for unix sockets
Also, don't attempt to remove the socket after
closing the server. It's not necessary to do this
anyway; the Close method on the socket already
does an Unlink if not using the abstract namespace
(hence tests don't need changing).

Abstract namespaces are a Linux extension, so this
will break if we ever support Unixes. I've purposely
not checked the OS, so that it'll be obviously broken
and we can revisit this issue.