state: modify InjectMachine to take params struct
InjectMachine is modified to take AddMachineParams
struct, and that struct's fields are now all
exported. Additionally, AddMachineWithConstraints
(which already took the params struct), now checks
that the formerly non-exported fields are blank
upon entry.

These changes are to accommodate manual provisioning,
where we want to inject a machine into state, but
set the nonce to something other than the bootstrap
nonce. Additionally, we are likely to add provider
type into the machine document soon.