Fix bug #1216285.
We were trying to print out a version.Number.Major as a string, but what
we really wanted was to print out the whole version.Number object (as a string).

In the process, I added the ability to 'easily' check our log for what messages
are written, so we can ensure they are useful and well formed.