Implement 'juju get' via the API.
The actual CLI change is pretty trivial. This also removes
state/statecmd/get.go and moves the implementation and tests
over to state/apiserver/client.

The hardest part is that the API doesn't preserve numeric
types (an int64(0) becomes a float64(0) because of JSON).
Hopefully we don't really care, but it can be hard to debug
{"foo": 0} != {"foo": 0}