Implement an Upgrader.DesiredVersion function.
This does some of the work that Upgrader.Tools does (determine the version
of the tools we want an agent using), but without having to search the
Provider storage to actually find a URL for those tools.
It also doesn't specifically require the version.Binary information
(namely, the platform and arch) though we may want to have those
for future implementations that don't ask things to upgrade until
we actually have tools for that platform.