environs/sshstorage: revert storage layout
In a lapse of sanity, I changed sshstorage's
layout so that the temporary directory was
contained within the designated storage
directory. This works fine in the context of
sshstorage, but not so well with how sshstorage
is intended to be used.

For bootstrapping the null provider, sshstorage
is used to operate on a specified remote
directory; later, when the machine agent is up,
it takes ownership of that directory and serves
it via environs/localstorage. Thus, both
sshstorage and localstorage must agree on the
layout.

Now, instead of having storage/{tmp,content},
the user can specify a separate temporary
directory. If left blank, storagedir+".tmp" will
be used.