Environ.SanityCheckConstraints -> environs/Preflighter
I've removed SanityCheckConstraints from Environ, and
introduced a new interface, Preflighter. An Environ
may implement this to have an opportunity to perform
"preflight checks" on series and constraints.

Preflighter contains one method, Preflight. This method
is provided with the series, constraints, and, if an
attempt to create a container is underway, a non-nil
Instance.