Rename Preflighter to Prechecker
Also, split Preflight method into two: PrecheckCreateMachine,
and PrecheckCreateContainer. The instance.Instance parameter
is only present in PrecheckCreateContainer, and is expected
to be non-nil.

The checks for cons.Container in PrecheckCreateContainer are
now gone, as they're expected to be non-nil and != instance.NONE.
Only if an environment had a problem with a specific type of
container would they care about the value.