environs/sshstorage: sane default tmpdir
TL;DR: previous change doesn't work with
manual bootstrap/null provider. The only
directories we can guarantee acccess to
are those created during initialisation
(before NewSSHStorage returns).

Longer explanation follows.

The previous changes to using a temporary
directory were not sane. NewSSHStorage
creates an initial connection with sudo,
in which the connecting user is assigned
ownership of the storage directory.
Subsequently, another session is executed
without sudo to perform file transfers.

The non-sudo session is what was creating
temporary directories. This didn't work
with the default (storagedir+".tmp") if
the parent directory isn't writable (which
is typical).

The new behaviour is:
 - If UseDefaultTmpDir is specified, then
   the storage directory is the tmpdir.
 - If a non-empty directory is specified,
   then the directory is created and
   ownership assigned to the login user.
   The directory is never removed.