environs/httpstorage: authenticated Put/Remove*
This change introduces two new functions: ServeTLS,
and ClientTLS. These functions are similar to Serve
and Client, but use TLS. A valid client certificate
is required for Put and Remove* methods to be invoked.

There's more changes required to get this to work
properly, to set the Subject appropriately in the
certificates.