provider/null: define simplestreams tools source
This is somewhat involved, as there's some fixes
required (and one drive-by in sshstorage). The
change to use GetToolsSources is simple enough,
and matches what the local provider does.

The Environ override struct in cmd/juju/bootstrap.go
had to so that the optional Environ interfaces were
still implemented. Overriding like that loses the
additional interfaces.

To do the above, environs.BootstrapStorager has been
changed to have a method that tells the environ to
use bootstrap storage, rather than returning a
bootstrap storage. The null provider has changed to
match, of course, as well as cmd/juju.