provider/azure: remove default daily stream
The simplestreams.GetMetadata function was
being called with three datasources: storage,
daily and released. In our tests, storage was
empty, but daily and released were not; daily
did not, however, have any matching image ids.
There's an issue with simplestreams.GetMetadata
not attempting the next datasource if the
metadata index is valid, but does not contain
any matching product IDs.

Since the behaviour of simplestreams.GetMetadata
is potentially useful for private clouds, we
will just remove the daily stream from the defaults
for now. If people really want it, they can get
at it with image-stream and image-metadata-url.

Fixes #1233924