environs/cloudinit: use cloud-tools pocket on precise
Use the cloud-tools pocket of the cloud archive on
12.04 (presumably we'll update for later LTS'). Also,
remove the PPA for precise, as cloud-tools has a recent
enough version of mongodb-server for our needs.

I've added cloud-tools to apt not just for state
servers, but so it's picked up by all; thus we'll get
up-to-date LXC for when we support containers on
non-local providers.

The OpenStack pockets are not added, as there are
separate ones for separate releases. It'll have to
be up to the charms to add that, I think.

Fixes #1234127