provider/null: dynamic default-series
default-series is used to decide which
tools to bootstrap with. This is not
necessarily valid for the null provider,
where the boostrap host's architecture
should instead decide which tools to
locate.

So now we assign default-series dynamically,
by performing detection at Prepare time.
We store the detected series and hardware
characteristics, and use them to avoid
repeating the detection at Bootstrap time.

Fixes #1236691