Providers are responsible for selecting tools
This change makes providers responsible for
selecting bootstrap tools, and ensuring they
are available in storage. The null provider is
updated to select the tools appropriate to the
bootstrap host. Other providers use a default
implementation in environs/bootstrap, which
has been extracted from cmd/juju/bootstrap.go.

There's an incidental change in environs/bootstrap
and provider/common: EnsureNotBootstrapped is
moved to provider/common to avoid a circular
dependency. It now lives alongside the *State*
methods and types in provider/common, which is
more appropriate given that it uses them.

One test fails due to a logging change. This is
fixed by https://codereview.appspot.com/14430064/.

Fixes #1236691