Migrate juju destroy-environment to use API
(This is a prototype, hence lacking tests.
Please review as such, and if it's in the
right direction I'll update with tests.)

This change introduces a new client API
method, DestroyJuju. This method (server-side)
marks the environment as Dying, destroys (and
waits on death of) all units, and then marks
the environment as Dead. When the environment
is marked as Dying, addition of units, services
and machines is prevented.

Machine agents watch, via the machiner API, for
changes to the environment's lifecycle, and
terminate themselves when the environment
becomes Dead.

A followup will update agent termination logic
to uninstall mongo on state servers, and also
remove the data-dir/log-dir.

Fixes #1218688