state: update SetEnvironConfig to take new and old
SetEnvironConfig is updated to take both new and old
configurations, and a delta computed from these. The
only functional change for now is that attributes that
exist in old but not in new are now deleted from state.

To do this right, we should also check that the
settings in state are the same as the "old" config,
so the delta computed is what the user expects. I have
not done this part yet, as it is a fairly big change
and I am not confident I understand all of the
repercussions.