Lots of refactoring of the test suite.
Move 'environ_test.go' into 'environ_whitebox_test.go' because it really is
a whitebox test (as is the rest of the _test.go files, but at least I fixed
the one I was touching).
Then move all of the new tests into a new environ_test.go so that we can have
it as a blackbox test.
Create an export_test.go for the stuff we do need exposed for testing.
Get all the tests passing again, now as blackbox tests.