Update provider/common.Bootstrap
Now use instance.Addresses, and attempt
all addresses in parallel. Instead of
using net.Dial, go directly to ssh. Since
private addresses might point to a user's
own local network and not the cloud's
private network, we generate a file in
cloud-init that contains the machine's
nonce, and verify it in the initial SSH
connection.

The nonce verification also serves to
synchronise cloud-init and sshinit. Since
the file is written as the last thing
cloud-init does, we can safely run
sshinit once it has been verified.