environs/manual: add ubuntu user
Manual bootstrap and provisioning now add the
ubuntu user (if it doesn't exist), update its
authorized_keys, and enables passwordless sudo
for it. This is somewhat invasive, so we will
have to call it out clearly in the documentation.

All ssh-related things other than the user
initialisation can now do away with the special
logic for handling sudo interactions.