provider/null: implement Environ.Destroy
A new worker is introduced: worker/terminationworker.
This worker simply waits for a signal (SIGABRT), and
then exits with ErrTerminateAgent. This causes the
agent to uninstall itself. We do not use SIGTERM, as
that would cause "service stop jujud-machine-N" to
uninstall the agent.

The null/manual provider has been updated at the same
time to implement Environ.Destroy. The Destroy method
simply connects to the bootstrap machine via SSH, and
issues "sudo pkill -SIGABRT jujud". We only do this
for machine 0, as Destroy will never be reached until
all non-manager machines are destroyed.

Fixes #1238948