ProvisioningScript client API
This is a new client API that, given the ID and nonce
of a machine injected into state, returns a script
that can be executed on that machine to install and
configure a Juju machine agent.

The existing MachineConfig client API call has been
modified: it was taking a series and arch. Series is
guaranteed to be in state.Machine; the arch is not,
but we make it a requirement for calling MachineConfig.
In the only place where MachineConfig is already called,
this is true. ProvisioningScript is expected to be
used similarly.