provider/local: always use filestorage on CLI
The local provider is modified such that all
storage interactions from the CLI go directly
to disk, and the provider code no longer starts
an httpstorage listener. To achieve this, we
no longer store bootstrap-ip in the .jenv, but
instead add it to the environment config during
bootstrap. Thus, we can use the presence of the
attribute to decide whether or not to use the
file or http storage. BootstrapStorager no longer
needs to be implemented, so EnableBootsrapStorage
has been eliminated from the implementation.

For old environments, bootstrap-ip will be set
in the .jenv; this is fine, as long as the
machine agent is running. There's an edge case
that will no longer work, where the environment
is bootstrapped but machine-0 agent is not
running. In this case, storage is inaccessible.

There are some other drive-by fixes to tests
that improve isolation, and some redundant tests
relating to the old sudo behaviour are removed.
All tests now pass with or without a local provider
running on the machine.

Fixes lp:1271672