Pass environs.BootstrapContext to Prepare
This is part one of fixing lp:1279259.
In this CL, we pass context to Prepare,
giving providers an opportunity to prompt
the user to initialise an environment.

Passing context was previously restricted to
Bootstrap and EnableBootstrapStorage; this is
insufficient, as there are many ways to prepare
an environment. It is currently possible to
successfully "Prepare" and environment, but
for the result to be unusable in the manual
provider (e.g. if the bootstrap-host address is
invalid). We should that a successfully Prepared
environment should be able to be Destroyed
(with force).

The next CL will remove BootstrapStorager and fix
lp:1279259 by initialising the ubuntu user during
Prepare.