provider/local: Destroy partial environments
Destroy is improved so that it does not assume
a complete/working environment. It will now
handle the scenario where bootstrap failed to
gain sudo rights. It will also remove the
~/.juju/local directory as a final step.

Additionally, fixed Destroy to stop passing all
of os.Args through to the sudo call, and construct
the call explicitly.

Fixes lp:1273295
Fixes lp:1281863