Bug #1278213, use a .tmp dir inside the target directory.
Rather than using a sibling directory which we may not be able to create (foo.tmp vs foo/.tmp)
we create a directory inside the target directory. We'll still clean it up when we are done,
but we make sure that it doesn't show up in listings, etc.

Also, all callers were passing UseDefaultTmpDir, so we can just get rid of the parameter,
since it wasn't actually useful.