provider/azure: numerous fixes
If all instances in a cloud service
are to be stopped, just delete the
service.

Use DeleteHostedService instead of
DestroyHostedService. DeleteHostedService
has been updated to pass the appropriate
option to delete associated disks/blobs,
which is what DestroyHostedService was
doing manually.

Don't bother with the parallel.Do code.
It complicates tests, and can't actually
be turned on without breaking Azure.