provider/local: remove mongo service on Destroy
The local provider's Destroy method is modified
to stop and remove the mongodb service we create.

If bootstrap fails, it currently leaves behind
the mongo service. Destroy normally removes the
service by killing jujud with SIGABRT, which
triggers it to clean up; this only works if
bootstrap passes, and the machine agent starts
up successfully.

Since the mongo service may be stopped/removed
by jujud itself, we do not check the error of
stopping/removing the service.

Fixes lp:1292275