Introduce unset-environment subcommand
A new juju subcommand, unset-environment (with unset-env alias)
is introduced. This subcommand will take the existing environment
config, remove the specified keys, and update it in state. If
the specified keys have associated defaults, then they will be
assigned.

This is necessary for removing configuration for which an empty
value is not permitted, e.g. for string attributes which are
omitted by default, such as http-proxy.

Fixes lp:1295372