Introduce InstanceDistributor policy
This is a policy that is invoked whenever
we attempt to assign a unit to an empty
machine.

The policy will be invoked with the
associated instances (i.e. provisioned,
empty machines) and the distribution group
instances (i.e. all other instances which
have assigned units of the same type.)

Azure will return nil, whereas ec2 will
want to check which availability zone each
machine is in and distribute for maximum
distance.