Change EnsureAvailability to match the existing servers.
Bug #1311089 has to do with the fact that if you 'juju bootstrap' and get a Trusty host,
and then you follow that up with 'juju ensure-availability' you end up with precise backup
machines. This is because it picks the PreferredSeries which is internally coded as 'precise'
if you don't have distro-info available.
However, it seems to make more sense that we should be preferring whatever the existing
server is running, rather than just the default in the environment.