Fix bug #1301353. Get the test suite to find juju-mongodb.
On trusty we are likely to only have juju-mongodb installed, which is at
/usr/lib/juju/bin/mongod and not in PATH. Find it if it is available and
use it.
Also for the store tests, invert the search, because we would rather use the
system mongod that comes with v8 support (so we don't have to skip tests).
This should also address bug #1304770, since it will allow us to detect
when we are using the 'wrong' version and disable those tests automatically
rather than requiring someone to set an env var.