replicaset: clean up, extend timeout
Various cleanups, factoring out code, etc.;
uncovered a bit of code that was leaking
an *mgo.Session.

I've extended the attempt strategy total
for replset modifications to 2 minutes,
since the mongo docs say that operations
inducing a renegotiation of primary may
incur downtime of >1m.

I was able to reproduce the error initially
on my machine by running the test in a loop;
I have not reproduced it with these changes.

Fixes lp:1321461