Port login-env-urls over to git.
we lose all the history of the changes, but at least the patch is here.
I haven't run the test suite yet to ensure that all the changes are good,
but at least the patch applies.

This changes Login to return an ENV UUID and allows us to connect
to the API server under '/ENVIRONUUID/api' style urls.