Spread units across availability zones
Common code for ec2 and openstack, and any
other provider that supports availability
zones similarly.

Two top-level functions are added to
provider/common: BestAvailabilityZones and
DistributeInstances. The former returns the
availability zones with the fewest instances
in the specified group. The latter returns
the set of instances that may be added to
the specified group without reducing the
availability zone spread optimality.