provider/azure: don't fail Destroy if AG/vnet can't be destroyed
When destroying an environment, it is common for the Azure API
to not release references to the vnet until a significant amount
of time after destroying all related VMs. Thus, destroy-environment
will often fail.

This PR changes the Azure provider to only warn if the vnet or
affinity group cannot be destroyed. Bootstrap is modified to use
the vnet/AG if they are preexisting. If the vnet/AG are being
destroyed while bootstrap is called, then bootstrap will either
create a reference to them thus preventing their deletion, or will
fail and the user will have to bootstrap again.

Fixes https://bugs.launchpad.net/bugs/1324910