Second round of review feedback.
Be a bit more careful about transforming no-such-method into a CallNotImplementedError.
Update the 'we see that you gave a bad type' test to also see that 'nil, err' is
not treated as a bad type, but instead just treated as the regular 'you got an error
calling this method' that we would expect.