Possible renaming of interfaces.
Get rid of the BestAPIVersioner as an interface.

Change the root api.State.Call into api.State.APICall and change the facade
wrapper to use CallFacade. The latter will be renamed to FacadeCall in the
next patch.
CallFacade doesn't fit very well tone wise, and doesn't follow the FooCall
pattern.