provider/azure: use Location in vnet creation; always delete AG
The Azure docs recommend using Location when creating a virtual
network site, so we do that (with a prerequisite gwacl branch).
Also, now that the vnet is not tied to an affinity group, we can
attempt to delete the affinity group even if deleting the virtual
network failed.

Fixes https://bugs.launchpad.net/juju-core/+bug/1259947