Track down the last few that weren't using FacadeCaller.
The Watcher code itself can't use FacadeCaller, because it wants to pass an Id.
We could change the FacadeCaller code, but I don't think its worth it.
It should keep it easier to reason about how the code is flowing that all FacadeCall
are not tied to an id.

This also gets us away from the proliferation of func call() methods on every
client-side Facade implementation.