utils/ssh: set ServerAliveInterval option
Bootstrapping Azure was seeing the CLI
hang at "apt-get upgrade" with no further
output or command completion. Inspection
of cloud-init-output.log on the server
showed that the bootstrap procedure
completed successfully.

The problem appears to be due to the
"apt-get upgrade" taking a long time on
Azure: over 6 minutes. This inactivity
causes the ssh server to discontinue sending
output to the client. To resolve the issue
we need to ensure the ssh client sends
keepalive messages to the server.

Fixes https://bugs.launchpad.net/juju-core/+bug/1316185