cmd/jujud: ensure we don't ensureMongoServer twice
The upgrade-steps and StateWorker both call ensureMongoServer,
which is meant to check/set a flag so that we only set it up
once. There's an early exit case which meant that the flag was
not being set, which led to the upgrade-steps code restarting
mongo half-way through.

Fixes https://bugs.launchpad.net/juju-core/+bug/1338511