Fixes lp:1339976, lp:1270466
lp:1270466
Add the following cmd to add-machine:

juju add-machine ssh:user@10.10.0.3 -i ~/myid.pub

where -i is a specific identity key used to ssh into the machine. The key is also stored in the machine for future logins.

lp:1339976
This PR also fixes a bug where cloudinit was not using the ssh client.