Replace juju.NewAPIConn with juju.NewAPIState.
This addresses bug #1326802. We don't want APIConn around, because it just
existed to keep track of an Environ which we don't want. We still needed a way
to open an api.State object from an Environ object in several places (NewAPIClientFromName
et al all use an environment *name*, but that doesn't work when we just created the
Environ details and want to open it.)

I also renamed 'apiconn_test.go' to 'api_test.go' because there is no APIConn anymore
so it doesn't make sense to have a test file named after it.