action-get in jujuc
Hooks can now run the command action-get to retrieve the map of params
passed with the action hook.  In the case of a nil params map, an empty
value will be passed.  Format for the command is as follows:

action-get [--output=<json|yaml>] [key[.inner.value...]]

In the case of a nonexistent key, a nil interface will be returned to
the command Stdout writer, which will display it as an empty string or
null in the case of JSON output.