WIP: testing/jujuconn: bifurcate
This is an attempt to undo the Lovecraftian horror that is JujuConnSuite.

This change introduces two types, JujuClientConnSuite and JujuServerConnSuite who are then combined to produces JujuConnSuite.

On the surface this may not appear to be a win, but hopefully over time the tests which rely on the Client behavior can move to JujuClientConnSuite, and the ones requireing Server behavior can move to the other.

Also, redo Reset() in terms of TearDownTest/SetUpTest so we can stop using hand rolled cleanup logic.