environs: remove Name method from environs.Environ
This change removes the Name method from the Environ interface.

The rational for this comes from the work at the Identity sprint which has identified that the canonical identifier for an environment is its UUID. This downplays the importance of the Environ.Name() method, so the method has been removed and replaced by the same method availble on Environ.Config(). As these two _MUST_ be the same, the effect on the caller is nil.

It should be noted that this change identified several places where Environ.Name() and Environ.Config().Name() were not reporting the same values and these have been fixed.