environs/config: introduce config.UUID()
Introduce `config.UUID()` method to access the uuid of the environment.

A uuid is the canonical indentifier for an environment, and superceeds the environment's name.

For backwards compatability, `config.UUID()` is currently permitted to return "", but once all environments are converted to have an environment uuid in place, this method will change to return a valid uuid for a valid config.