worker/upgrader: allow downgrades for aborted upgrades
In some cases upgrades will need to be aborted and the best course of
action will be to restart in to the previous agent version. This
revision modifies the upgrader worker so that it will allow downgrades
to the pre-upgrade tools version, but only if an upgrade is still in
progress. This allows the upgrade-steps worker to trigger a downgrade
to the previous agent version by changing the environment's
agent-version.