Don't restart mongo every time jujud starts.
Add a public method to let us know if the service exists as expected.
Check to see if the service already exists before writing it out again.
Refactor the upstart exists to use once source, and fix the tests so it doesn't create a braindead empty upstart file first.
Start the service if we need to.