provider/manual: add manual-provider defaults to prepared config
Prepare calls "validate", which was not updating the config
with manual-provider specific attributes (Validate calls validate
and updates the config too). This meant that the manual-provider
config attribute defaults were not being populated in the .jenv
file, so we were not using SSH storage from the client when we
should have been.

Fixes https://bugs.launchpad.net/juju-core/+bug/1347715