cmd/jujud: downgrade agent-version if a state fails to come up for upgrade
If during an upgrade one or more state servers fail to signal it is
ready for upgrade, the master state server will modify the environment
agent version to trigger a downgrade to the previous agent version.

So that the upgrader worker will accept downgrades to the previous
agent version (it normally won't), upgradeWorkerContext's
IsUpgradeRunning method is now passed to the upgrader
worker. Downgrades to the previous agent version are allowed if an
upgrade is in progress.